+++ include/service/generation/vpg_property_accessor_generation_service.hpp
@@ -19,7 +19,6 @@ class VPGPropertyAccessorGenerationService
         
         static std::vector<std::wstring> GetPropertyAccessorTypeOrder();
 
-        static void GetIsHavingGenerateTypeMapType(const std::vector<std::shared_ptr<VPGEnumClassProperty>> &enumClassProperties, bool &isHavingGeneralType, bool &isHavingVectorType, bool &isHavingMapType);
         static void GenerateRead(const std::wstring &propertyName, const std::wstring &type, const std::vector<std::shared_ptr<VPGEnumClassProperty>> &enumClassPropertiesReadOnly, std::wstring &result);
         static void GenerateWrite(const std::wstring &propertyName, const std::wstring &type, const std::vector<std::shared_ptr<VPGEnumClassProperty>> &enumClassPropertiesWriteOnly, std::wstring &result);
         static void GenerateInsert(const std::wstring &propertyName, const std::wstring &type, const std::vector<std::shared_ptr<VPGEnumClassProperty>> &enumClassPropertiesWriteOnly, std::wstring &result);