+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 - Review function naming
 - Git
+- Seperate Remove and RemoveObject (function using void * as argument will remove object property and shift pointer. i.e. cannot find object from vector)
+- Add ObjectId to IObject for idenitity checking. (Cannot find object from vector as pointer is shifted 8 for some cases)
 
 ### [v0.3.0] - 2025-01-26: Review - Review Property Accessor
 - Class macro VECTOR, VECTOR_SPTR add Remove method